Create BudgetCalc.py
Added classes:
Day, Month, Year, and Product
Added functionality:
Days of each month can be represented either as the span of dates within the month with year.calendar, or as the amount spent on each day as year.moneyCalendar. This can be altered with the function year.spent(), which takes a date and a price.